livekit-agent/pyproject.toml
[tool.poetry]
name = "livekit_agent"
version = "0.1.0"
description = ""
authors = ["Lucheng_Hugh <hughqing@qq.com>"]
readme = "README.md"
packages = [
    { include = "livekit_agent", from = "." }
]

[tool.poetry.dependencies]
python = "^3.12"
livekit-plugins-deepgram = "^0.6.4"
livekit-plugins-openai = "^0.8.0"
livekit-plugins-silero = "^0.6.3"
python-dotenv = "^1.0.1"
langgraph = "^0.2.4"
langgraph-sdk = "^0.1.27"
matplotlib = "^3.9.2"
nest-asyncio = "^1.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
